Creating DoMain and Infrastucture
