+++ InteractiveAtlas/InteractiveAtlas.Domain/Entities/TouristAttraction.cs
@@ -2,18 +2,20 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Globalization;
 
-namespace InteractiveAtlas.Entities
+namespace InteractiveAtlas.Domain.Entities
 {
-    [Table("TypicalProducts")]
-    public class TypicalProduct
+    [Table("TouristAttractions")]
+    public class TouristAttraction
     {
-        [Column("ProductId")]
+        [Column("AttractionId")]
         public int Id { get; set; }
         [StringLength(50)]
-        public string Name { get; set; } = null!; 
+        public string Name { get; set; } = null!;
 
         public string? Description { get; set; }
         [StringLength(250)]
+        public string? Location { get; set; }
+        [StringLength(400)]
         public string? ImageUrl { get; set; }
 
         //Navigation
