+++ InteractiveAtlas/InteractiveAtlas.Domain/Entities/TypicalProduct.cs
@@ -2,20 +2,18 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Globalization;
 
-namespace InteractiveAtlas.Entities
+namespace InteractiveAtlas.Domain.Entities
 {
-    [Table("TouristAttractions")]
-    public class TouristAttraction
+    [Table("TypicalProducts")]
+    public class TypicalProduct
     {
-        [Column("AttractionId")]
+        [Column("ProductId")]
         public int Id { get; set; }
         [StringLength(50)]
-        public string Name { get; set; } = null!;
+        public string Name { get; set; } = null!; 
 
         public string? Description { get; set; }
         [StringLength(250)]
-        public string? Location { get; set; }
-        [StringLength(400)]
         public string? ImageUrl { get; set; }
 
         //Navigation
